requirements.txt
streamlit
pillow
python-dotenv
agno
google-genai
tavily-python
